Revert "Refactor"
