--- Assets/Scripts/Hero/HeroData.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8707e74a758a4e4daa4c413067e6f886
-timeCreated: 1680640274
