--- Assets/Scripts/Hero/HeroMovement.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: b517957fb450488fbfbbd49c4e4439d7
-timeCreated: 1680640259
