+++ Assets/Scripts/Hero/PlayerBrain.cs
@@ -33,8 +33,9 @@ namespace Hero
             foreach (var inputSource in _inputSources)
             {
                 if (inputSource.Direction == Vector2.zero)
+                {
                     continue;
-                
+                }
                 
                 return inputSource.Direction;
             }
